Move CodeDom CSharp helpers to Common
