+++ src/System.Private.Xml/src/System/Xml/Xsl/XslException.cs
@@ -2,12 +2,12 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System.CodeDom.Compiler;
 using System.Diagnostics;
 using System.Globalization;
 using System.Resources;
 using System.Runtime.Serialization;
 using System.Text;
+using System.Xml.Xsl.Xslt;
 
 namespace System.Xml.Xsl
 {
